requirements.txt
absl-py==2.1.0
accelerate==0.31.0
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
archspec @ file:///croot/archspec_1697725767277/work
async-timeout==4.0.3
attrs==23.2.0
auto_gptq==0.7.1
autoawq_kernels==0.0.6
beautifulsoup4==4.12.3
bitsandbytes==0.43.1
blessed==1.20.0
blinker==1.7.0
boltons @ file:///work/ci_py311/boltons_1677685195580/work
Brotli @ file:///work/ci_py311/brotli-split_1676830125088/work
cachetools==5.3.2
causal-conv1d==1.3.0.post1
cbor==1.0.0
certifi @ file:///croot/certifi_1707229174982/work/certifi
cffi @ file:///croot/cffi_1700254295673/work
chardet==5.2.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
cloudpickle==3.0.0
cmake==3.29.2
coloredlogs==15.0.1
conda-content-trust @ file:///croot/conda-content-trust_1693490622020/work
conda-package-handling @ file:///croot/conda-package-handling_1690999929514/work
conda_package_streaming @ file:///croot/conda-package-streaming_1690987966409/work
contourpy==1.2.0
cryptography @ file:///croot/cryptography_1702070282333/work
cycler==0.12.1
datasets==2.17.0
deepspeed==0.13.1
dill==0.3.8
diskcache==5.6.3
distro @ file:///croot/distro_1701455004953/work
docker==7.0.0
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.7.0
entrypoints==0.4
evaluate==0.4.1
exceptiongroup==1.2.0
faiss==1.8.0
fastapi==0.110.3
filelock==3.13.1
flash-attn==2.5.3
Flask==3.0.2
fonttools==4.48.1
frozenlist==1.4.1
fsspec==2023.10.0
gekko==1.1.1
gitdb==4.0.11
GitPython==3.1.41
google-api-core==2.17.0
google-api-python-client==2.117.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
gpustat==1.1.1
greenlet==3.0.3
grpcio==1.60.1
gunicorn==21.2.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.23.3
humanfriendly==10.0
idna @ file:///work/ci_py311/idna_1676822698822/work
ijson==3.3.0
importlib-metadata==7.0.1
inscriptis==2.5.0
interegular==0.3.3
ir_datasets==0.5.8
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lark==1.1.9
libmambapy @ file:///croot/mamba-split_1698782620632/work/libmambapy
llvmlite==0.42.0
lm-format-enforcer==0.9.8
lxml==5.3.0
lz4==4.3.3
Mako==1.3.2
mamba-ssm==2.2.2
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
mdurl==0.1.2
menuinst @ file:///croot/menuinst_1702390294373/work
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mlflow==2.10.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
nltk==3.8.1
numba==0.59.1
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp311-cp311-linux_x86_64.whl#sha256=5f96f274d410a1682519282ae769c877d32fdbf171aa8badec7bf5e1d3a1748a
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.133
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
openai==1.12.0
optimum==1.16.2
outlines==0.0.34
packaging @ file:///croot/packaging_1693575174725/work
pandas==2.2.0
pathlib==1.0.1
peft==0.13.0
pillow==10.2.0
platformdirs @ file:///croot/platformdirs_1692205439124/work
pluggy @ file:///work/ci_py311/pluggy_1676822818071/work
prometheus_client==0.20.0
protobuf==3.20.0
psutil==5.9.8
py-cpuinfo==9.0.0
py-spy==0.3.14
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycosat @ file:///croot/pycosat_1696536503704/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.17.2
pynvml==11.5.0
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.1.1
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.4
PyYAML==6.0.1
querystring-parser==1.2.4
ray==2.12.0
referencing==0.35.0
regex==2023.12.25
requests @ file:///croot/requests_1690400202158/work
responses==0.18.0
rich==13.7.1
rouge==1.0.1
rouge-score==0.1.2
rpds-py==0.18.0
rsa==4.9
ruamel.yaml @ file:///work/ci_py311/ruamel.yaml_1676838772170/work
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
sentence-transformers==3.0.1
sentencepiece==0.1.99
sentry-sdk==1.40.4
setproctitle==1.3.3
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.6
SQLAlchemy==2.0.27
sqlparse==0.4.4
starlette==0.37.2
style==1.1.0
sympy==1.12
tensorboard==2.16.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tf-keras==2.15.0
threadpoolctl==3.2.0
tiktoken==0.6.0
tokenizers==0.19.1
torch==2.2.1
torchvision==0.17.0
tqdm @ file:///croot/tqdm_1679561862951/work
transformers==4.44.2
trec-car-tools==2.6
triton==2.2.0
trl==0.8.1
truststore @ file:///croot/truststore_1695244293384/work
typing_extensions==4.9.0
tyro==0.7.3
tzdata==2023.4
unlzw3==0.2.2
update==0.0.1
uritemplate==4.1.1
urllib3 @ file:///croot/urllib3_1698257533958/work
uvicorn==0.29.0
uvloop==0.19.0
vaderSentiment==3.3.2
vllm==0.4.1
vllm-nccl-cu12==2.18.1.0.4.0
wandb==0.16.3
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.1
xformers==0.0.25
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
zlib-state==0.1.9
zstandard @ file:///work/ci_py311_2/zstandard_1679339489613/work